[DDW-459] Support URL inside the locale files
