+++ source/renderer/app/config/urlsConfig.js
@@ -6,6 +6,3 @@ export const TESTNET_EXPLORER_URL = 'http://cardano-explorer.cardano-testnet.ioh
 export const DEVELOPMENT_EKG_URL = 'http://localhost:8083';
 export const STAGING_EKG_URL = 'http://localhost:8082';
 export const TESTNET_EKG_URL = 'http://localhost:8081';
-
-export const SUPPORT_URL_EN = 'https://iohk.zendesk.com/hc/en-us/categories/360000877653-Daedalus-wallet-mainnet';
-export const SUPPORT_URL_JP = 'https://iohk.zendesk.com/hc/en-us/categories/360000877653-Daedalus-wallet-mainnet';
