+++ source/renderer/app/containers/settings/categories/SupportSettingsPage.js
@@ -1,16 +1,29 @@
 // @flow
 import React, { Component } from 'react';
 import { observer, inject } from 'mobx-react';
+import { defineMessages, intlShape } from 'react-intl';
 import { remote } from 'electron';
 import SupportSettings from '../../../components/settings/categories/SupportSettings';
 import type { InjectedProps } from '../../../types/injectedPropsType';
 import { generateFileNameWithTimestamp } from '../../../../../common/fileName';
-import getSupportUrl from '../../../utils/getSupportUrl';
+import { getSupportUrl } from '../../../utils/network';
+
+const messages = defineMessages({
+  supportRequestLinkUrl: {
+    id: 'settings.support.reportProblem.linkUrl',
+    defaultMessage: '!!!https://iohk.zendesk.com/hc/en-us/categories/360000877653-Daedalus-wallet-mainnet',
+    description: '"Support request" link URL in the "Report a problem" section on the support settings page.',
+  },
+});
 
 const shell = require('electron').shell;
 
 @inject('stores', 'actions') @observer
 export default class SupportSettingsPage extends Component<InjectedProps> {
+
+  static contextTypes = {
+    intl: intlShape.isRequired,
+  };
 
   static defaultProps = { actions: null, stores: null };
 
@@ -20,8 +33,9 @@ export default class SupportSettingsPage extends Component<InjectedProps> {
   };
 
   handleSupportRequestClick = (event: MouseEvent) => {
-    const locale = this.props.stores.profile.currentLocale;
-    this.handleExternalLinkClick(event, getSupportUrl(locale));
+    const { intl } = this.context;
+    const supportRequestLinkUrl = intl.formatMessage(messages.supportRequestLinkUrl);
+    this.handleExternalLinkClick(event, getSupportUrl(supportRequestLinkUrl));
   };
 
   handleDownloadLogs = () => {
