+++ source/renderer/app/i18n/locales/defaultMessages.json
@@ -4816,17 +4816,50 @@
         "description": "Message \"Loading wallet data\" on the loading screen.",
         "end": {
           "column": 3,
-          "line": 18
+          "line": 19
         },
         "file": "source/renderer/app/containers/LoadingPage.js",
         "id": "loading.screen.loadingWalletData",
         "start": {
           "column": 21,
-          "line": 14
+          "line": 15
+        }
+      },
+      {
+        "defaultMessage": "!!!https://iohk.zendesk.com/hc/en-us/categories/360000877653-Daedalus-wallet-mainnet",
+        "description": "Link to Open Support page",
+        "end": {
+          "column": 3,
+          "line": 24
+        },
+        "file": "source/renderer/app/containers/LoadingPage.js",
+        "id": "loading.screen.reportIssue.reportIssueButtonUrl",
+        "start": {
+          "column": 24,
+          "line": 20
         }
       }
     ],
     "path": "source/renderer/app/containers/LoadingPage.json"
+  },
+  {
+    "descriptors": [
+      {
+        "defaultMessage": "!!!https://iohk.zendesk.com/hc/en-us/categories/360000877653-Daedalus-wallet-mainnet",
+        "description": "\"Support request\" link URL in the \"Report a problem\" section on the support settings page.",
+        "end": {
+          "column": 3,
+          "line": 16
+        },
+        "file": "source/renderer/app/containers/settings/categories/SupportSettingsPage.js",
+        "id": "settings.support.reportProblem.linkUrl",
+        "start": {
+          "column": 25,
+          "line": 12
+        }
+      }
+    ],
+    "path": "source/renderer/app/containers/settings/categories/SupportSettingsPage.json"
   },
   {
     "descriptors": [
