+++ source/renderer/app/i18n/locales/hr-HR.json
@@ -63,6 +63,7 @@
   "loading.screen.reportIssue.buttonLabel": "!!!Open support ticket",
   "loading.screen.reportIssue.connecting.text": "!!!Having trouble connecting to network?",
   "loading.screen.reportIssue.downloadLogsLinkLabel": "!!!Download logs",
+  "loading.screen.reportIssue.reportIssueButtonUrl": "!!!https://iohk.zendesk.com/hc/en-us/categories/360000877653-Daedalus-wallet-mainnet",
   "loading.screen.reportIssue.syncing.text": "!!!Having trouble syncing?",
   "loading.screen.startingCardanoMessage": "!!!Starting Cardano node",
   "loading.screen.stoppedCardanoMessage": "!!!Cardano node stopped",
@@ -147,6 +148,7 @@
   "settings.support.logs.title": "!!!Logs",
   "settings.support.reportProblem.content": "If the FAQ does not contain a solution for the issue you are experiencing, please use <a href=\"#\">problem reporting feature</a> to report the issue you are experiencing.",
   "settings.support.reportProblem.link": "Support request",
+  "settings.support.reportProblem.linkUrl": "!!!https://iohk.zendesk.com/hc/en-us/categories/360000877653-Daedalus-wallet-mainnet",
   "settings.support.reportProblem.title": "!!!Reporting a problem",
   "sidebar.wallets.addWallet": "Dodaj novčanik",
   "static.about.buildInfo": "!!!MacOS build 3769, with Cardano 1.0.4",
