+++ source/renderer/app/i18n/locales/ja-JP.json
@@ -63,6 +63,7 @@
   "loading.screen.reportIssue.buttonLabel": "サポートチケットを作成する",
   "loading.screen.reportIssue.connecting.text": "接続に問題がありますか？",
   "loading.screen.reportIssue.downloadLogsLinkLabel": "ログをダウンロードする",
+  "loading.screen.reportIssue.reportIssueButtonUrl": "!!!https://iohk.zendesk.com/hc/en-us/categories/360000877653-Daedalus-wallet-mainnet",
   "loading.screen.reportIssue.syncing.text": "同期に問題がありますか?",
   "loading.screen.startingCardanoMessage": "カルダノ・ノードを起動中",
   "loading.screen.stoppedCardanoMessage": "カルダノ・ノードが停止しました",
@@ -147,6 +148,7 @@
   "settings.support.logs.title": "ログファイル",
   "settings.support.reportProblem.content": "もしよくある質問を参照しても症状が改善されない場合には、アプリケーション内にあるサ {supportRequestLink} を利用してください。",
   "settings.support.reportProblem.link": "ポートリクエスト機能",
+  "settings.support.reportProblem.linkUrl": "!!!https://iohk.zendesk.com/hc/en-us/categories/360000877653-Daedalus-wallet-mainnet",
   "settings.support.reportProblem.title": "問題を報告する",
   "sidebar.wallets.addWallet": "ウォレット追加",
   "static.about.buildInfo": "{platform} ビルド <strong>{build}</strong>, {apiName} <strong>{apiVersion}</strong>",
